crosscluster/logical: separate create-table, row, and txn Resume paths
Previously, the LDR job's Resume method had a monolithic ingest()
function that handled both the create-table offline initial scan and
steady-state row ingestion.

This change splits resume into per-workflow files so that:
1. The create-table workflow (offline initial scan + publish) runs to
   completion first.
2. The job then forks into mode-specific ingestion (row-level vs
   transactional).
3. Each workflow lives in its own file with a clear entry point.

The change should have no user visible behavior changes. It does have a
subtle change to how ldr create table is implemented. Previously, create
table was interleaved with ingest, so once creating the table was
finished an error was returned which triggered a retry. Now that it has
its own resume logic, it returns nil and the execution continues
normally.

Epic: CRDB-57649
Release note: None
