+++ pkg/crosscluster/logical/BUILD.bazel
@@ -14,6 +14,9 @@ go_library(
         "metrics.go",
         "offline_initial_scan_processor.go",
         "purgatory.go",
+        "resume_create_table.go",
+        "resume_row.go",
+        "resume_txn.go",
         "savepoint.go",
         "sql_crud_writer.go",
         "table_batch_handler.go",
